Moved Iced app code into app.rs
Also made other minor changes
Commented and documented most functions
Fixed all the compiler warnings
